requirements.txt
discord-py-interactions
aiohttp
aiomysql
mysql-connector-python
